Connect write review feature to firebase

+++ app/src/main/java/hk/ust/cse/comp4521/group20/opentoiletandroid/ReviewViewHolder.java
@@ -31,5 +31,5 @@ public class ReviewViewHolder extends RecyclerView.ViewHolder{
         mDescField.setText(desc);
     }
 
-    public void setRating(int rating) {mRatingBar.setRating(rating);}
+    public void setRating(float rating) {mRatingBar.setRating(rating);}
 }
